-Added documentation -Partially optimised -broke down files for easier reading and maintance
