+++ src/main.cpp
@@ -1,17 +1,17 @@
 #include <stdio.h>
 #include <curl/curl.h>
 
-#include <CurlManager.hpp>
-
+#include <WebPage.hpp>
 
 
 int main() {
 
 
-    WebPage* webpage = curl_manager.get("https://en.wikipedia.org/wiki/James_Joyce");
+    WebPage* webpage = new WebPage("https://en.wikipedia.org/wiki/James_Joyce");
 
     if(webpage != nullptr) {
         webpage->scrape();
+        webpage->write_markdown();
         delete webpage;
     }
